refactor(plugins): adopt output data to models
