+++ examples/cli-e2e/mocks/config.mock.js
@@ -1,9 +1,11 @@
-import eslintPlugin from '@quality-metrics/eslint-plugin';
-import lighthousePlugin from '@quality-metrics/lighthouse-plugin';
+//import eslintPlugin from '@quality-metrics/eslint-plugin';
+//import lighthousePlugin from '@quality-metrics/lighthouse-plugin';
 
 export default {
+  persist: { outputPath: 'cli-config-out.json' },
+  categories: [],
   plugins: [
-    eslintPlugin({ config: '.eslintrc.json' }),
-    lighthousePlugin({ config: '.lighthouserc.json' }),
+   // eslintPlugin({ config: '.eslintrc.json' }),
+   // lighthousePlugin({ config: '.lighthouserc.json' }),
   ],
 };
