+++ packages/cli/src/bin.ts
@@ -1,6 +1,6 @@
 #! /usr/bin/env node
-import {cli} from "./index";
-import {hideBin} from "yargs/helpers";
+import { cli } from './index';
+import { hideBin } from 'yargs/helpers';
 
 // bootstrap yargs; format arguments
 cli(hideBin(process.argv)).argv;
