+++ packages/cli/src/index.ts
@@ -1,21 +1,17 @@
-import {yargsCli} from "./lib/cli";
-import {yargsGlobalOptionsDefinition} from "./lib/options";
-import {middlewares} from "./lib/middlewares";
-import {commands} from "./lib/commands";
+import { yargsCli } from './lib/cli';
+import { yargsGlobalOptionsDefinition } from './lib/options';
+import { middlewares } from './lib/middlewares';
+import { commands } from './lib/commands';
 
-export const cli = (args: string[]) => yargsCli(
-  // hide first 2 args from process
-  args,
-  {
-    usageMessage: 'CPU CLI',
-    scriptName: 'cpu',
-    options: yargsGlobalOptionsDefinition(),
-    middlewares,
-    commands
-  }
-);
-
-export {
-  CommandBase,
-  commandBaseSchema,
-} from './lib/implementation/base-command-config';
+export const cli = (args: string[]) =>
+  yargsCli(
+    // hide first 2 args from process
+    args,
+    {
+      usageMessage: 'CPU CLI',
+      scriptName: 'cpu',
+      options: yargsGlobalOptionsDefinition(),
+      middlewares,
+      commands,
+    },
+  );
